+++ KnowYourTurf.Core/Domain/DomainEntity.cs
@@ -2,10 +2,9 @@ using CC.Core.Domain;
 
 namespace KnowYourTurf.Core.Domain
 {
-
     public abstract class DomainEntity : Entity
     {
-        public virtual long CompanyId { get; set; }
+        public virtual int CompanyId { get; set; }
     }
 
 }
