+++ KnowYourTurf.Core/Domain/Vendor.cs
@@ -69,7 +69,7 @@ namespace KnowYourTurf.Core.Domain
 
         #endregion
 
-        public virtual bool PurchaseOrderHasBeenCompleted(long purchaseOrderId)
+        public virtual bool PurchaseOrderHasBeenCompleted(int purchaseOrderId)
         {
             var purchaseOrder = GetCompletedPurchaseOrders().FirstOrDefault(x => x.EntityId == purchaseOrderId);
             foreach (var li in purchaseOrder.LineItems)
@@ -78,19 +78,19 @@ namespace KnowYourTurf.Core.Domain
         }
 
 
-        public virtual double AmountOfSubTotalForPurchaseOrder(long purchaseOrderId)
+        public virtual double AmountOfSubTotalForPurchaseOrder(int purchaseOrderId)
         {
             var purchaseOrder = GetCompletedPurchaseOrders().FirstOrDefault(x => x.EntityId == purchaseOrderId);
             return purchaseOrder.LineItems.Sum(x => x.SubTotal.Value);
         }
 
-        public virtual double TotalTaxForPurchaseOrder(long purchaseOrderId)
+        public virtual double TotalTaxForPurchaseOrder(int purchaseOrderId)
         {
             var purchaseOrder = GetCompletedPurchaseOrders().FirstOrDefault(x => x.EntityId == purchaseOrderId);
             return purchaseOrder.LineItems.Sum(x => x.Tax.Value);
         }
 
-        public virtual double TotalAmountDueForPurchaseOrder(long purchaseOrderId)
+        public virtual double TotalAmountDueForPurchaseOrder(int purchaseOrderId)
         {
             var purchaseOrder = GetCompletedPurchaseOrders().FirstOrDefault(x => x.EntityId == purchaseOrderId);
             return purchaseOrder.LineItems.Sum(x => x.SubTotal.Value + x.Tax.Value);
